Refractor Collectable interface and improve asynchronicity
Refactored the Collectable interface and improved asynchronous operations in the 'simbot-common-core' project. Removed the Job import and replaced it with Async in the Collectables.kt file. Added a new method 'isFlowCollectable()' to check the type of the Collectable. Replaced 'launchAndCollect()' method with 'collectAsync()'. Also added descriptive comments for each method to improve readability. Now the 'collectAsync()' method will be returning a completed Async Unit. All these changes will improve the overall efficiency and performance of the application.