Add Flow.converts to Iterator support in simbot-common-core library
Refactored the simbot-common-core library to support the conversion of Flow into an Iterator. Changes were made to build.gradle.kts to include a new dependency 'simbot-common-suspend-runner'. Additional helper method `asIterator()` was added to Collections.kt to convert Flow into Iterator for improved handling of coroutine results. Also added overloads for `asSynchronouslyIterateCollectable()` to the Collectables.jvm.kt to process a Collectable object and transform it into an instance of SynchronouslyIterateCollectable for better collection operations using coroutines. This allows improved handling of asynchronous flows in the application.